+++ src/main/java/com/example/backend/domain/redis/listener/AuthenticationEventListener.java
@@ -9,7 +9,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
-import com.example.backend.domain.security.adapter.CustomUserDetails;
+import com.example.backend.domain.auth.security.CustomUserDetails;
 import com.example.backend.domain.redis.service.ActiveUserService;
 
 @Component
